requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-admin-sortable2==2.3.1
django-appconf==1.0.6
django-ckeditor==6.7.2
django-imagekit==5.0.0
django-js-asset==2.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
packaging==25.0
pilkit==3.0
pillow==11.0.0
python-telegram-bot==21.8
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
django-modeltranslation~=0.19.11